Finally understood the origin of weird nullptr access exception on stream closure. Don't fuckin iterate on raw pointers...

+++ FDN/include/FDN.h
@@ -17,7 +17,7 @@ namespace MyFDN
 
 	std::vector<std::complex<float>> SimpleFFT_FFT(const std::vector<float>& input);
 
-	std::vector<float> IDFT(const std::vector<std::complex<float>>& input, const size_t duration);
+	std::vector<float> IDFT(const std::vector<std::complex<float>>& input, const float duration);
 
 	std::vector<float> SimpleFFT_IFFT(const std::vector<std::complex<float>>& input);
 }
